+++ src/main/kotlin/xenakis/ui/SettingsPane.kt
@@ -6,6 +6,7 @@ import hextant.fx.hbox
 import javafx.scene.control.Label
 import javafx.scene.layout.VBox
 import xenakis.impl.Knob
+import xenakis.impl.toDecimal
 import xenakis.model.KnobControl
 import xenakis.model.Settings
 import xenakis.sc.NumericalControlSpec
@@ -25,13 +26,13 @@ class SettingsPane(settings: Settings, context: Context) : VBox(5.0) {
                     +hbox(
                         Knob(
                             "Latency (sclang)", KnobControl(settings.scLangLatency),
-                            NumericalControlSpec(0.1, 0.01, 1.0, 0.01), context = context
+                            NumericalControlSpec(0.1, 0.01, 1.0, 0.01.toDecimal()), context = context
                         )
                     )
                     +hbox(
                         Knob(
                             "Latency (scsynth)", KnobControl(settings.serverLatency),
-                            NumericalControlSpec(0.1, 0.01, 1.0, 0.01), context = context
+                            NumericalControlSpec(0.1, 0.01, 1.0, 0.01.toDecimal()), context = context
                         )
                     )
                 }
